Add options to remove & purge deleted messages
New flag --remove-deleted controls the soft delete part of a full backup separately from the --fast-incremental flag. Deleted messages will be marked as deleted in the database and the EML files moved to the deleted folder.

New flag --purge-deleted removes deleted messages from the database, but leaves the deleted message copy on disk.